+++ app/assets/stylesheets/_intake.scss
@@ -50,7 +50,7 @@
 .cf-non-rated-issue {
   display: grid;
   grid-column-gap: 10px;
-  grid-template-columns: auto auto 76px;
+  grid-template-columns: auto auto;
   padding-bottom: 1em;
 
   div {
