+++ client/app/intake/actions/appeal.js
@@ -1,5 +1,6 @@
 import { ACTIONS, ENDPOINT_NAMES } from '../constants';
 import ApiUtil from '../../util/ApiUtil';
+import { formatRatingData } from '../util';
 import { formatDateStringForApi } from '../../util/DateUtil';
 import _ from 'lodash';
 
@@ -66,31 +67,7 @@ export const completeIntake = (intakeId, appeal) => (dispatch) => {
     meta: { analytics }
   });
 
-  const ratingData = {
-    request_issues:
-      _(appeal.ratings).
-        map((rating) => {
-          return _.map(rating.issues, (issue) => {
-            return _.merge(issue, { profile_date: rating.profile_date });
-          });
-        }).
-        flatten().
-        filter('isSelected')
-  };
-
-  const nonRatingData = {
-    request_issues:
-      _.map(appeal.nonRatedIssues, (issue) => {
-        return {
-          decision_text: issue.issueDescription,
-          issue_category: issue.issueCategory
-        };
-      })
-  };
-
-  const data = {
-    request_issues: _.concat(ratingData.request_issues.value(), nonRatingData.request_issues)
-  };
+  const data = formatRatingData(appeal);
 
   return ApiUtil.patch(`/intake/${intakeId}/complete`, { data }, ENDPOINT_NAMES.COMPLETE_INTAKE).
     then(
