+++ db/migrate/20180530210916_add_issue_category_to_request_issues.rb
@@ -3,5 +3,6 @@ class AddIssueCategoryToRequestIssues < ActiveRecord::Migration[5.1]
     add_column :request_issues, :issue_category, :string
     change_column_null :request_issues, :rating_issue_reference_id, true
     change_column_null :request_issues, :rating_issue_profile_date, true
+    change_column_null :request_issues, :description, true
   end
 end
